Solve sequence 3 strand 3
This required fixing a bug with finding potential new parents, where if a parent had one child, the child had to be normal
and the potential new child also had to be normal. It was changed so that the existing child and the potential new child
both have to be different than Doubler, so that it's possible to correctly find new parents for Extender cells.