Fix infinite freeze when changing some levels
When changing levels e.g. from 4.3 to 5.1, the app would freeze. Internally, it was looping between levels 5.2 and 5.1. The reason was setting range of strand spin button,
which unexpectedly changed the current strand by lowering the current number to the current max, so from 7 to 2. Swapping the places of setting the range and setting
the current strand, as well as checking if these operation are needed, solved the problem.